+++ src/onEmit.ts
@@ -12,24 +12,37 @@ const addLine = (lines: string[], indent: number, text: string): void => {
     lines.push(line(indent, text));
 }
 
-function getRandomItem<T>(array: readonly T[]): T {
-    const randomIndex = Math.floor(Math.random() * array.length);
-    return array[randomIndex];
+const shuffleArray = <T>(array: readonly T[]): T[] => {
+    const shuffledArray = [...array];
+    for (let i = shuffledArray.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [shuffledArray[i], shuffledArray[j]] = [shuffledArray[j], shuffledArray[i]];
+    }
+    return shuffledArray;
 }
 
-const emitValue = (context: EmitContext, property: ModelProperty, value: Value, indent: number): string => {
+function getRandomItem<T>(array: readonly T[], count: number = 1): T {
+    const shuffledArray = shuffleArray(array);
+    const selectedItems = shuffledArray.slice(0, count);
+    return selectedItems as T;
+}
+
+const emitValue = (context: EmitContext, property: ModelProperty, value: Value, indent: number, options: Options): string => {
     let ret = "";
     switch (value.valueKind) {
         case "NullValue":
             return "null";
-        case "StringValue":
         case "NumericValue":
+            return value.value.toString();
+        case "StringValue":
             return `"${value.value}"`;
+        case "BooleanValue":
+            return value.value ? "true" : "false";
         case "ArrayValue":
             if (value.values.length > 0) {
                 ret = "[\n";
                 ret += value.values.map((v, index, arr) => {
-                    return line(indent + 1, emitValue(context, property, v, indent + 1));
+                    return line(indent + 1, emitValue(context, property, v, indent + 1, options));
                 }).join(",\n");
                 ret = ret + "\n";
                 ret += line(indent, "]");
@@ -40,22 +53,42 @@ const emitValue = (context: EmitContext, property: ModelProperty, value: Value, 
                 const examples = getExamples(context.program, model);
                 if (examples.length > 0) {
                     ret = "[\n";
-                    const ex = getRandomItem(examples);
-
-                    ret += line(indent + 1, `${emitValue(context, property, ex.value, indent + 1)}\n`);
-
+                    const exs = (options.randomize ? shuffleArray(examples) : examples).slice(0, options.arraySize);
+                    ret += exs.map((ex, index) => {
+                        return line(indent + 1, `${emitValue(context, property, ex.value, indent + 1, options)}`);
+                    }).join(",\n");
+                    ret = ret + "\n";
                     ret += line(indent, "]");
                 } else {
-                    ret = "[]";
+                    // check derived models
+                    if (context.options.fillArrays) {
+                        const derivedModels = model.derivedModels;
+                        const examples = derivedModels.map(m => getExamples(context.program, m)).flat();
+                        if (examples.length > 0) {
+                            ret = "[\n";
+                            const exs = (options.randomize ? shuffleArray(examples) : examples).slice(0, options.arraySize);
+                            ret += exs.map((ex, index) => {
+                                return line(indent + 1, `${emitValue(context, property, ex.value, indent + 1, options)}`);
+                            }).join(",\n");
+                            ret = ret + "\n";
+                            ret += line(indent, "]");
+
+                        } else {
+                            ret = "[]";
+                        }
+                    }
                 }
             }
+            if (ret.length === 0)
+                ret = "[]";
+
             return ret;
             break;
         case "ObjectValue":
 
             ret = "{\n";
             value.properties.forEach(v => {
-                ret += line(indent + 1, `"${v.name}": ${emitValue(context, property, v.value, indent + 1)},\n`);
+                ret += line(indent + 1, `"${v.name}": ${emitValue(context, property, v.value, indent + 1, options)},\n`);
             });
 
             // remove the last comma
@@ -82,18 +115,20 @@ const emitSampleValue = (context: EmitContext, model: Model, property: ModelProp
 
     const propEx = getExamples(context.program, property);
     if (propEx.length > 0) {
-        const ex = getRandomItem(propEx);
-        return emitValue(context, property, ex.value, indent);
+        const exs = (options.randomize ? shuffleArray(propEx) : propEx).slice(0, 1);
+        const ex = exs[0];
+        return emitValue(context, property, ex.value, indent, options);
     }
 
     const modelEx = getExamples(context.program, model);
     if (modelEx.length > 0) {
-        const ex = getRandomItem(modelEx);
+        const exs = (options.randomize ? shuffleArray(modelEx) : modelEx).slice(0, 1);
+        const ex = exs[0];
         switch (ex.value.valueKind) {
             case "ObjectValue":
                 const p = (ex.value as ObjectValue).properties.get(property.name);
                 if (p) {
-                    return emitValue(context, property, p.value, indent);
+                    return emitValue(context, property, p.value, indent, options);
                 }
                 break;
             default:
@@ -109,22 +144,26 @@ const emitSampleValue = (context: EmitContext, model: Model, property: ModelProp
             const model = (property.type as ArrayModelType).indexer.value as Model;
             const examples = getExamples(context.program, model);
             if (examples.length > 0) {
-                const ex = getRandomItem(examples);
+                const exs = (options.randomize ? shuffleArray(examples) : examples).slice(0, options.arraySize);
                 let ret = "[\n";
-                ret += line(indent + 1, emitValue(context, property, ex.value, indent + 1));
-                ret += "\n";
+                ret += exs.map((ex, index) => {
+                    return line(indent + 1, `${emitValue(context, property, ex.value, indent + 1, options)}`);
+                }).join(",\n");
+                ret = ret + "\n";
                 ret = ret + line(indent, "]");
                 return ret;
             }
         } else {
             const modelEx = getExamples(context.program, property.type);
             if (modelEx.length > 0) {
-                const ex = getRandomItem(modelEx);
+
+                const exs = (options.randomize ? shuffleArray(modelEx) : modelEx).slice(0, 1);
+                const ex = exs[0];
                 switch (ex.value.valueKind) {
                     case "ObjectValue":
                         let ret = line(indent, "{\n");
                         ex.value.properties.forEach(v => {
-                            ret += line(indent + 1, `"${v.name}": ${emitValue(context, property, v.value, indent + 1)},\n`);
+                            ret += line(indent + 1, `"${v.name}": ${emitValue(context, property, v.value, indent + 1, options)},\n`);
                         });
 
                         ret = ret + line(indent, "\n}");
@@ -158,7 +197,7 @@ const emitSample = (context: EmitContext, model: Model, options: Options): { lin
         model.baseModel.properties.forEach((prop: ModelProperty) => {
             const val = emitSampleValue(context, model, prop, options, indent + 1);
             if (val !== undefined) {
-                addLine(lines, indent + 1, `"${prop.name}": ${val}},`);
+                addLine(lines, indent + 1, `"${prop.name}": ${val},`);
             }
         });
     }
@@ -187,7 +226,10 @@ export async function $onEmit(context: EmitContext) {
             models: context.options["models"],
             outDir: context.options["outDir"],
             namespace: context.options["namespace"],
-            setUndefinedToNull: context.options["setUndefinedToNull"] ?? true
+            setUndefinedToNull: context.options["setUndefinedToNull"] ?? true,
+            fillArrays: context.options["fillArrays"] ?? false,
+            randomize: context.options["randomize"] ?? false,
+            arraySize: context.options["arraySize"] ?? 1
         }
 
         const outfiles: { [key: string]: string[] } = {};